reorganise cli app
