+++ notedown/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import
 from .notedown import *
-from .main import cli, markdown_template, __version__
+from .main import markdown_template, __version__
